IGNITE-17349 Add common UI components
Now everything that we are going to print should go through
UI component. The list of UI components decide how to render
errors, tables and more.

ValidationProblem is removed because it is impossible to
understand from the http status code to what class body
should be deserialized: Problem or ValidationProblem (both
of them have 400 http code). So, now there is only one
Problem that in case of validation issue provides additional
information in `invalidParams` field.

And Problem now does not include the trace id and error code
into the `datail` field because it makes it hard to render
on client side. The detail field should contain useful
explanation of what happened and how to fix this but not
a debug information.

Also, removed all dots from commands output and description.