IGNITE-17090 Show user-friendly message in sql command
Previously sql command showed an abstract error message like
`SQL query parsing error` and in was hard to understand what
was wrong. Now it shows the root cause of the issue.

The format of the error message is still being different from
time to time, and we have to think about the error message
formatting in sql-engine.