IGNITE-18148 Add more dynamic completers
The architecture of dynamic completers were
refactored. Now it is more flexible and the
lifecycle of completers is defined in factories.
So, if some data for suggestions (say, node list)
should be changed then the factory will create
new DynamicCompleter with updated data inside.

The Options class is introduced. It collects all
options that are used in the CLI in one place.
This gives us two advantages: DRY and the ability
to use Options class in the completion strategies.
For example, cluster init --cluster-name the
option that should be completed only once. This
information could be placed in Options class in
the future.

To demonstrate the advantages of refactoring I've
added two more completions to interactive mode:
connect command and cluster init.