IGNITE-18187 Check java version in start script
In case the user has java version 8 it is hard to understand why
the application fails to start. Now all distributions check java
version is 11 or higher.

Note, that script does not fail if the version is less than 11 because
the user could have installed some custom version that the script
could not parse. It would be strange to fail the application in
this case. So, it just displays the message.