IGNITE-18280 Fix dependencies in Gradle packaging tasks (#1442)
* Fix dependencies to get rid of Gradle warnings.
* Do not assembly release artifacts (such as .NET binaries) during regular build.

This fix properly declares dependencies between packaging tasks so that Gradle can optimize them correctly and don't complain.
Also this adds a property check for adding release artifacts. The reason is that adding an artifact to the configuration automatically adds a dependency to the `assemble` task so running `gradlew build` always produces artifacts intended only for the release.