IGNITE-17333 .NET: Add SQL ResultSet object mapping (#1584)
Add `ISql.ExecuteAsync<T>` API, which maps SQL result set data to user types.

Before that, we could insert and retrieve data in a strongly typed way with object mapping via `IRecordView<T>` API, but SQL results could only be accessed in a weakly typed way through `IgniteTuple` or `DbDataReader`. This change fills the void.

Or, to put it another way, we could use LINQ to query data in a strongly typed manner. But not every query can be expressed in LINQ, sometimes we have to use raw SQL, but result mapping is still very useful, which is now possible.

This is also the fastest way to retrieve query results (see benchmarks), because there is no LINQ translation, but compiled delegates read the data with the lowest possible overhead.

Additional fix: use `rollbackAsync` when cleaning `ClientResourceRegistry` to avoid blocking Netty worker thread.