IGNITE-19840 .NET: Reload schema when unmapped columns are detected (#2454)
When unmapped columns are detected in a Tuple or a POCO, we must ensure that the latest schema is used before throwing an exception.

For example, right after `ALTER TABLE ADD COLUMN`, the user expects to be able to insert a tuple with the new column. However, the client still has the old schema in the cache, and previously would throw unmapped column error. Now we retry the operation with forced schema refresh.