IGNITE-20864 Choose primary replica for colocated jobs (#3043)
* Now the primary replica is chosen as a worker for
collocated execution.

* The failover is adjusted to support the selection of
the next worker node, see NextWorkerSelector.

* InteractiveJobs class is introduced. This is helpful
API for making assertions and sending commands to
interactive jobs.

* Before this change ComputeJobFailover was listening for
physical topology changes. But it is not what we really want.
Instead, we want to listen to a logical topology changes. So,
now NodeLeftEventsSource is not needed anymore because
LogicalTopologyEventListener is enough.