IGNITE-21651 Fix ItSslClientHandlerTest on Java 21
Our tests use Netty tools to generate self-signed certificates. These tools use OpenJDK internal classes to do it with Reflection to invoke their methods. In Java 21, one of the needed methods (set()) does not exist anymore, so OpenJDK's classes cannot be used anymore.

This commit switches the certificate generation logic in tests to BouncyCastle by adding it to the classpath.