IGNITE-21645 Fix tests extending AbstractPlannerTest on Java 21 (#3330)
The tests change a final field value. In such a situation, the corresponding IgniteTestUtils method changes the corresponding Field's modifiers to remove the final flag from them. The modifiers are changed through a MethodHandle, and on highest Java versions, a MethodHandle of a final Field now throws an UnsupportedOperationException.

It turns out, this modification of the modifiers is actually not needed on any of the currently LTS Java versions (11, 17, 21), so we just remove this in this commit.