IGNITE-22478 Use single argument in Compute API (#3926)
The Compute API is changed:

- Only single argument can be accepted or null by any compute job.
- Marshaler API is introduced as a part of public API and can be used in Compute API.
- Default marshaller is implemented but not used explicitly. By default we support only
primitives. If the one want to use custom marshaler they should declare it explicitly.

Marshalers allow us to define the default way to serialize objects: java serialization.
At the same time client can implement theirs own marshalers and use them in any
public API that supports it. This is the first implementation and it is supposed to be
improved in the future. The only thing that will not be changed -- Marshaler and
Compute interfaces (their public parts).