IGNITE-22942 Add ability to StaticNodeFinder to resolve multiple IP addresses under the same host name (#4320)
Now, if a name (not an address) is provided in the original list, it's being resolved to addresses. If it resolves to none, it just contributes nothing to the output. If it resolves to more than one address, all of them are included in the result.

This will allow `StaticNodeFinder` to be more useful in Kubernetes environments.