IGNITE-23220 Avoid concurrent Catalog writes to Metastorage on the same node (#4406)
Each DDL tries to make a write to the Metastorage. Each write is assigned a sequential number (version); if a few DDLs try to write to the Metastorage, they will all compete for the same version, and only one of them can succeed. This competition is wasteful. This commit avoids such concurrency on the same node by explicitly linearizing writes to the Metastorage from the Catalog before doing them.