IGNITE-23273 Use a ConversionService per Ignite node (#4451)
By default, Micronaut uses a shared ConversionService, so, if we start a few Ignite nodes in the same JVM (like we do in our tests) and then we stop a node, the shared ConversionService gets reset and loses converters still needed for remaining nodes. This commit switches to using one ConversionService instance per Ignite node to avoid this interference.