Fix some issues reported by undefined behavior sanitizer
The old calculation in `fits` could result in a signed integer overflow. The new version uses two comparisons combined with a logical and, but the compiler is smart enough to generate even better code than for the old version.

If append_bytes is called with an empty view memcpy could be called with a nullptr which ubsan complained about, even though we are not copying anything.