IGNITE-27346 Handle 'LWM below earliest Catalog ts' situation on recovery
A few components access Catalog by the stored LWM on node recovery. LWM is the local property of the node, while Catalog history might be compacted by other nodes, so local LWM might point at a moment of time at which the Catalog history is already compacted out. To avoid failing in such situations, we take max(lwm, earliestCatalogTs) on recovery to access the Catalog history.